Add styling to navigation bar
Made header div and the unordered lists within it flex
containers. Set the header's main axis to space-between
and the cross-axis to center. Gave it a background color
of #1F2937. Set the ul list style type to none to remove
bullet points. Removed the default margin and padding on
the ul's. Set the logo ul to font size 24px and color value
of #F9FAF8 and the links font size to 18px and color value
of #E5E7EB. Set the top and bottom padding for both to 24px
and pushed each further from the edge of the window by
assigning 300px of padding to the left and right sides.