Add content below navigation bar
Renamed the navigation bar div class to header-navbar
for clarity. Added div below it with class header-content.
Put a div inside that to contain text that will go on the
and button left side of this section with class main-content.
Added an image tag to go on the right side.