Modify HTML and CSS
Added padding to .logo, .links, .header-content, and
.main-image to better match desired result. Removed
the align-content from .header-content as it was not
affecting anything. Changed justify-content to
space-between to space things better. Removed the
bottom margin from .main-text to push it closer to
.main-subtext. Similar removed top and bottom margins
from .main-subtext. Set width of it to 465px to format
the amount of lines it wraps to. Set the min-width and
height of the image to match desired outcome. Added
align-self to center to better position the img element.