Add quotation section
Added div with class quote with quotation and credit as two
p tags. They have class main-quote and credit respectively.
Made .quote a column flex container, aligned the quote to
center and put the credit at flex-end and with align-self;
put 580px of padding on the right to push to be still
under the quote. Added 100px padding to top and bottom
of .quote. Gave the main-quote p tag 800px in width to
wrap text properly. Added link tag in head of HTML to
make Roboto font function with font-weight correctly.