Add Call to Action section
Add CTA bubble in a div with class cta, within that have a
div with class cta-text to contain two p tags for the text.
Below put a button with class cta-button. Made .cta a flex
container and cta-text a column flex container. Centered
.cta, made sure it doesn't shrink, applied background color.
Put 60px of top/bottom padding, put a gap of 100px between
the text and the button. Put 120px of top/bottom margin and
400px of left/right margin to create the centered effect.
Set the border radius to 10px for rounded corners. Set
.main-cta and .sub-cta, the two p tags, to have 0 margin
to put them closer together. Set the font-size and color
for each. Copied the CSS for the first button but edited
the border tag so it was a solid 2px white border.